made HostItem able to create ServerConnectionOptions
